requirements-test.txt
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-html>=3.0.0
selenium>=4.0.0
webdriver-manager>=3.0.0
pytest-xdist>=2.0.0
pytest-flask>=1.2.0
pytest-mock>=3.0.0
fpdf2>=2.7.5
xhtml2pdf>=0.2.12
reportlab>=4.0.0
html5lib>=1.1
